+++ src/hybridmap/HybridNeutralAtomMapper.cpp
@@ -10,7 +10,6 @@
 
 #include "hybridmap/HybridNeutralAtomMapper.hpp"
 
-#include "circuit_optimizer/CircuitOptimizer.hpp"
 #include "datastructures/CircuitOptimizations.hpp"
 #include "hybridmap/Mapping.hpp"
 #include "hybridmap/MoveToAodConverter.hpp"
@@ -47,7 +46,7 @@ namespace na {
 void NeutralAtomMapper::mapAppend(qc::QuantumComputation& qc,
                                   const Mapping& initialMapping) {
   // remove barriers and measurements
-  qc::CircuitOptimizer::removeFinalMeasurements(qc);
+  qc.removeFinalMeasurements();
   // check if multi-qubit gates are present
   multiQubitGates = false;
   for (const auto& op : qc) {
@@ -70,9 +69,9 @@ void NeutralAtomMapper::mapAppend(qc::QuantumComputation& qc,
   }
 
   qmap::replaceMCXWithMCZ(qc);
-  qc::CircuitOptimizer::singleQubitGateFusion(qc);
-  qc::CircuitOptimizer::flattenOperations(qc);
-  qc::CircuitOptimizer::removeFinalMeasurements(qc);
+  qmap::singleQubitGateFusion(qc);
+  qc.flattenOperations();
+  qc.removeFinalMeasurements();
 
   const auto dag = constructDAG(qc);
 
@@ -173,8 +172,8 @@ qc::QuantumComputation NeutralAtomMapper::convertToAod() {
   // decompose bridge gates
   decomposeBridgeGates(mappedQc);
   qmap::replaceMCXWithMCZ(mappedQc);
-  qc::CircuitOptimizer::singleQubitGateFusion(mappedQc);
-  qc::CircuitOptimizer::flattenOperations(mappedQc);
+  qmap::singleQubitGateFusion(mappedQc);
+  mappedQc.flattenOperations();
   // decompose AOD moves
   MoveToAodConverter aodScheduler(*arch, hardwareQubits, flyingAncillas);
   mappedQcAOD = aodScheduler.schedule(mappedQc);