+++ src/sc/Mapper.cpp
@@ -10,7 +10,6 @@
 
 #include "sc/Mapper.hpp"
 
-#include "circuit_optimizer/CircuitOptimizer.hpp"
 #include "datastructures/CircuitOptimizations.hpp"
 #include "ir/Definitions.hpp"
 #include "ir/operations/CompoundOperation.hpp"
@@ -67,7 +66,7 @@ Mapper::Mapper(qc::QuantumComputation quantumComputation, Architecture& arch)
   // strip away qubits that are not used in the circuit
   qc.stripIdleQubits(true);
   // strip away final measurement gates
-  qc::CircuitOptimizer::removeFinalMeasurements(qc);
+  qc.removeFinalMeasurements();
 }
 
 void Mapper::processDisjointQubitLayer(